change type amount, fee to int

+++ lib/momo_vn.dart
@@ -128,8 +128,8 @@ class MomoPaymentInfo {
   String partnerCode;
   String merchantNameLabel;
 
-  double amount;
-  double fee;
+  int amount;
+  int fee;
   String? description;
   String? extra;
   String? username;
